--- src/main/java/Department.java
@@ -1,10 +1,0 @@
-import Employee.IEmployee;
-import enums.EBrand;
-
-import java.util.List;
-
-public class Department {
-    List<EBrand> brands;
-    List<Goods> goods;
-    List<IEmployee> consultants;
-}
