--- src/main/java/Employee/IEmployee.java
@@ -1,4 +1,0 @@
-package Employee;
-
-public class IEmployee implements Interface.IEmployee {
-}
