--- src/main/java/Goods.java
@@ -1,13 +1,0 @@
-import enums.EBrand;
-import enums.EDepartment;
-import enums.EGroup;
-
-public class Goods {
-    long uniqueID;
-    String name;
-    EBrand brand;
-    EGroup group;
-    EDepartment department;
-    double price;
-    String info;
-}
