+++ src/main/java/shop/Customer.java
@@ -1,0 +1,13 @@
+package shop;
+
+import com.sun.tools.javac.comp.Check;
+
+import java.util.List;
+
+public class Customer {
+    String name;
+    float discount;
+    double money;
+    List<Goods> bucket;
+    Check check;
+}
