+++ src/main/java/shop/Employee/IEmployee.java
@@ -1,0 +1,4 @@
+package shop.Employee;
+
+public class IEmployee implements shop.Interface.IEmployee {
+}
