+++ src/main/java/shop/Employee/StoreKeeper.java
@@ -1,4 +1,4 @@
-package Employee;
+package shop.Employee;
 
 public class StoreKeeper extends IEmployee {
 }
