+++ src/main/java/shop/GoodsWraper.java
@@ -1,3 +1,5 @@
+package shop;
+
 public class GoodsWraper {
     Goods goods;
     int count;
