+++ src/main/java/shop/Interface/ICustomer.java
@@ -1,0 +1,4 @@
+package shop.Interface;
+
+public interface ICustomer {
+}
