+++ src/main/java/shop/enums/EBrand.java
@@ -1,4 +1,4 @@
-package enums;
+package shop.enums;
 
 public enum EBrand {
