+++ src/main/java/shop/enums/EDepartment.java
@@ -1,4 +1,4 @@
-package enums;
+package shop.enums;
 
 public enum EDepartment {
